requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
gunicorn==21.0.1
pydantic==2.7.4
sqlalchemy==2.0.21
pymysql==1.1.0
aiomysql==0.2.0
alembic==1.11.2
pytest==7.4.2
bcrypt==4.0.1
python-multipart==0.0.6
joblib==1.3.2
pandas==2.1.2
boto3==1.28.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
openai==1.35.13
kiwipiepy==0.18.0
psycopg2-binary==2.9.9
langchain==0.2.11
langchain-core==0.2.23
langchain-community==0.2.10
langchain-text-splitters==0.2.2
langsmith==0.1.143
httpx==0.27.0